+++ runtime/migration_cache.go
@@ -1,0 +1,41 @@
+package runtime
+
+import (
+	"context"
+	"sync"
+
+	"github.com/hashicorp/golang-lru/simplelru"
+	"github.com/rilldata/rill/runtime/services/catalog"
+)
+
+type migrationMetaCache struct {
+	cache *simplelru.LRU
+	lock  sync.Mutex
+}
+
+func newMigrationMetaCache(size int) *migrationMetaCache {
+	cache, err := simplelru.NewLRU(size, nil)
+	if err != nil {
+		panic(err)
+	}
+
+	return &migrationMetaCache{cache: cache}
+}
+
+func (c *migrationMetaCache) get(instID string) *catalog.MigrationMeta {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+	if val, ok := c.cache.Get(instID); ok {
+		return val.(*catalog.MigrationMeta)
+	}
+
+	meta := catalog.NewMigrationMeta()
+	c.cache.Add(instID, meta)
+	return meta
+}
+
+func (c *migrationMetaCache) evict(ctx context.Context, instID string) {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+	c.cache.Remove(instID)
+}
